refactor(server): remove duplicate resolveOffset, update CLAUDE.md (#260)
Phase C: consolidate resolveOffset → resolveToElement alias, document
getElementText behavior changes in CLAUDE.md.

Two test files imported resolveOffset directly from document-model.js;
redirected to document.js where the alias now lives. The parity tests in
issue-377-position-diagnostics.test.ts remain — they now compare
resolveToElement to itself (trivially true) but still serve as regression
guards if the implementation ever diverges again.
