+++ CLAUDE.md
@@ -58,6 +58,7 @@ Full file-level detail: [docs/architecture.md](docs/architecture.md#file-map)
 - All document mutations go through the server's Y.Doc -> changes sync to editor via Hocuspocus
 - Annotations stored in Y.Map('annotations'), not in document content. `author` field: `"user" | "claude" | "import"` (import = Word comments from .docx files). User annotations are plain notes to Claude — `suggestedText` and `directedAt: "claude"` are Claude-only features created via MCP tools (`tandem_suggest`, `tandem_comment`), not the UI.
 - Three coordinate systems: flat text offsets (server, includes heading prefixes), ProseMirror positions (client, structural), Yjs RelativePositions (CRDT-anchored, survive edits). Modules: `src/server/positions.ts`, `src/client/positions.ts`, shared types in `src/shared/positions/`
+- `getElementText()` returns clean plain text (no markup tags) with `\n` separators between nested block children (list items, paragraphs within list items). `extractText()` additionally prepends heading prefixes and joins top-level elements with `\n`.
 - Multi-document: each file gets a documentId (hash of path) = Hocuspocus room name. All MCP tools accept optional `documentId`, defaulting to active document. `CTRL_ROOM` is reserved -- never use as a document ID. Server broadcasts `openDocuments` via Y.Map('documentMeta')
 - Communication: `tandem_checkInbox` (poll for user actions + chat) and `tandem_reply` (Claude's chat responses). **Call `tandem_checkInbox` between tasks.** `tandem_status` and `tandem_checkInbox` return `mode: "solo" | "tandem"` — adapt behavior accordingly (in Solo mode, hold annotations)
 - Solo/Tandem mode is stored in CTRL_ROOM's `Y_MAP_USER_AWARENESS` map under the `Y_MAP_MODE` key, not per-document. Mode changes broadcast to all open documents
@@ -103,6 +104,7 @@ Full file-level detail: [docs/architecture.md](docs/architecture.md#file-map)
 - **Hocuspocus replaces Y.Doc in `onLoadDocument`.** The `onDocSwapped` callback in `provider.ts` reattaches server event queue observers to the new instance. A runtime warning fires if the callback is missing. See #178 audit.
 - **Annotation-sync cleanups distinguish swap vs close.** `registerAnnotationObserver` returns `(phase?: "swap" | "close") => void`. A Y.Doc swap keeps the per-doc tombstone ledger alive so in-flight debounced writes still serialize tombstones; only `"close"` drops the ledger. `reattachObservers` passes `"swap"`; `clearFileSyncContext` / `setFileSyncContext`'s replace path pass `"close"`. See #333.
 - **Y.js "Invalid access" warnings** during session restore are harmless stderr noise.
+- **`getElementText()` strips inline marks and separates nested blocks.** The function uses `toDelta()` (not `toString()`) for XmlText content, emitting `\n` for embedded elements (hardBreaks). Child XmlElements (listItem, paragraph, blockquote) are separated by `\n`. This matches ProseMirror's block structure for correct flat-offset alignment.
 
 ### MCP / Server
 - **Channel shim uses low-level `Server`, not `McpServer`.** The Channels spec requires `import { Server } from '@modelcontextprotocol/sdk/server/index.js'` with explicit `setRequestHandler()` calls. The HTTP MCP server uses the high-level `McpServer` wrapper.
