+++ tests/server/issue-377-position-diagnostics.test.ts
@@ -62,7 +62,7 @@ import { fileURLToPath } from "url";
 import { afterEach, describe, expect, it } from "vitest";
 import * as Y from "yjs";
 import { loadMarkdown } from "../../src/server/file-io/markdown.js";
-import { extractText, resolveOffset } from "../../src/server/mcp/document-model.js";
+import { extractText, resolveOffset } from "../../src/server/mcp/document.js";
 import {
   anchoredRange,
   flatOffsetToRelPos,
